requirements.txt
pandas
google-cloud-bigquery
db-dtypes
ipykernel